pyproject.toml
[project]
name = "cmr-art"
version = "0.1.0"
description = "Pitures and ASCII arts of the Classic of Mountains and Rivers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "pillow>=11.3.0",
]
